requirements.txt
tensorflow-gpu==1.15
opencv-python
transforms3d
natsort
imutils
scikit-image
numpy
matplotlib
termcolor
scipy
